Skip parsing body for continuation packets, fix checksum for normal packets
Skip parsing body for continuation packets for now, as they have
a different structure, maybe the cmd name parsing could be added back,
but it seemed more work that it was worth.

Also fix the checksum position for normal packets, in USB communication
the body can be smaller than the packet size, so the checksum is not
always the last byte of the packet.